pyproject.toml
[project]
name = "webloginbrute"
version = "0.27.2"
description = "Web登录暴力破解工具"
authors = [{name = "RedteamNotes"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security",
    "Topic :: Internet :: WWW/HTTP",
]
dependencies = [
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "psutil>=5.9.0",
    "chardet>=5.0.0",
]

[project.scripts]
webloginbrute = "webloginbrute.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]

# 版本管理配置
[tool.hatch.build.targets.wheel]
packages = ["webloginbrute"]

[tool.hatch.build.hooks.custom]
path = "webloginbrute/version.py" 